Refactor symlinks into separae files
